requirements.txt
django==4.2
cryptography
djangorestframework
django-cors-headers
drf-yasg
```